refactor: Emotions enum을 domain 레이어로 이동
- Emotions enum을 data에서 domain 레이어로 재배치
