+++ app/src/main/java/com/stopstone/musicplaylist/data/model/entity/DailyTrack.kt
@@ -3,7 +3,7 @@ package com.stopstone.musicplaylist.data.model.entity
 import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
-import com.stopstone.musicplaylist.data.model.Emotions
+import com.stopstone.musicplaylist.domain.model.Emotions
 import com.stopstone.musicplaylist.ui.model.TrackUiState
 import java.util.Date
