+++ app/src/main/java/com/stopstone/musicplaylist/domain/repository/search/SearchRepository.kt
@@ -1,6 +1,6 @@
 package com.stopstone.musicplaylist.domain.repository.search
 
-import com.stopstone.musicplaylist.data.model.Emotions
+import com.stopstone.musicplaylist.domain.model.Emotions
 import com.stopstone.musicplaylist.data.model.response.Track
 
 interface SearchRepository {
