+++ app/src/main/java/com/stopstone/musicplaylist/domain/usecase/search/SaveDailyTrackUseCase.kt
@@ -1,6 +1,6 @@
 package com.stopstone.musicplaylist.domain.usecase.search
 
-import com.stopstone.musicplaylist.data.model.Emotions
+import com.stopstone.musicplaylist.domain.model.Emotions
 import com.stopstone.musicplaylist.data.model.entity.DailyTrack
 import com.stopstone.musicplaylist.ui.model.TrackUiState
 import com.stopstone.musicplaylist.domain.repository.common.TrackRepository
