+++ app/src/main/java/com/stopstone/musicplaylist/util/Converters.kt
@@ -3,7 +3,7 @@ package com.stopstone.musicplaylist.util
 import androidx.room.TypeConverter
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
-import com.stopstone.musicplaylist.data.model.Emotions
+import com.stopstone.musicplaylist.domain.model.Emotions
 import com.stopstone.musicplaylist.ui.model.TrackUiState
 import java.util.Date
